+++ src/app.ts
@@ -2,74 +2,20 @@
  * | - import -
  * | -----------------------------------------------------------------------------------
  */
-import fs from 'fs';
-import http from 'http';
-import https from 'https';
-import Redis from 'ioredis';
-import socketIO from 'socket.io';
 import * as env from './env';
-import autodeploySubscribeHandler from './psubscribe/autodeploy';
-import serverSubscribeHandler from './psubscribe/server';
+import Redis from './redis';
+import server from './server';
+import Socket from './socket';
 
 /* | -----------------------------------------------------------------------------------
  * | - Initialization -
  * | -----------------------------------------------------------------------------------
  */
-let server: http.Server|https.Server;
-const redis = new Redis();
+const socket = new Socket(server);
+socket.init();
 
-if (env.ssl) {
-  server = https.createServer({
-    cert: fs.readFileSync(env.sslCrt),
-    key: fs.readFileSync(env.sslKey),
-  });
-} else {
-  server = http.createServer();
-}
-
-const io = socketIO(server);
-
-/* | -----------------------------------------------------------------------------------
- * | - Listen all server.* event from Redis -
- * | -----------------------------------------------------------------------------------
- */
-redis.psubscribe('server.*');
-redis.psubscribe('autodeploy.*');
-redis.on('pmessage', (channel, pattern, message) => {
-  switch (channel) {
-    case 'server.*':
-      serverSubscribeHandler(message, io.sockets);
-      break;
-    case 'autodeploy.*':
-      autodeploySubscribeHandler(message, io);
-      break;
-    default:
-      console.warn('[PMessage] ->', channel);
-  }
-});
-
-/* | -----------------------------------------------------------------------------------
- * | - Socket.io events -
- * | -----------------------------------------------------------------------------------
- */
-io.on('connection', (socket) => {
-  socket.on('leave', (rooms) => {
-    if (Array.isArray(rooms)) {
-      rooms.forEach((room) => socket.leave(room));
-    } else {
-      socket.leave(rooms);
-    }
-  });
-
-  // Leave from all rooms
-  socket.on('logout', () => {
-    Object.keys(socket.rooms).forEach((room) => {
-      if (socket.id !== room) {
-        socket.leave(room);
-      }
-    });
-  });
-});
+const redis = new Redis(socket);
+redis.init();
 
 /* | -----------------------------------------------------------------------------------
  * | - Run the server -